requirement.txt
kivy
pycryptodome
cryptography
